+++ lib/sprockets/processor_utils.rb
@@ -118,8 +118,9 @@ module Sprockets
       Symbol,
       TrueClass,
       FalseClass,
-      NilClass
-    ] + (0.class == Integer ? [Integer] : [Bignum, Fixnum])).freeze
+      NilClass,
+      Integer
+    ]).freeze
 
     # Internal: Set of all nested compound metadata types that can nest values.
     VALID_METADATA_COMPOUND_TYPES = Set.new([